requirements.txt
pymetasploit3
python-nmap
pyshark